pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Victor Polyakov <polyakov.victor21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
djoser = "^2.2.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
